Separates utils exceptions.
